+++ GHEtool/main_class.py
@@ -36,16 +36,16 @@ class Borefield(BaseClass):
 
     HOURLY_LOAD_ARRAY: np.ndarray = np.arange(0, 8761, UPM).astype(np.uint32)
 
-    __slots__ = 'H', 'borehole',\
-                'number_of_boreholes', '_borefield', 'cost_investment', \
-                'Tf_max', 'Tf_min', 'limiting_quadrant', \
-                'Tf',\
-                '_ground_data', '_borefield_load',\
-                'results_peak_heating', 'options_pygfunction',\
-                'results_peak_cooling', 'results_month_cooling', 'results_month_heating', 'Tb', 'THRESHOLD_WARNING_SHALLOW_FIELD', \
-                'gui', 'D', 'r_b', 'gfunction_calculation_object',\
-                '_sizing_setup', \
-                '_secundary_borefield_load', '_building_load', '_external_load'
+    __slots__ = 'H', 'borehole', \
+        'number_of_boreholes', '_borefield', 'cost_investment', \
+        'Tf_max', 'Tf_min', 'limiting_quadrant', \
+        'Tf', \
+        '_ground_data', '_borefield_load', \
+        'results_peak_heating', 'options_pygfunction', \
+        'results_peak_cooling', 'results_month_cooling', 'results_month_heating', 'Tb', 'THRESHOLD_WARNING_SHALLOW_FIELD', \
+        'gui', 'D', 'r_b', 'gfunction_calculation_object', \
+        '_sizing_setup', \
+        '_secundary_borefield_load', '_building_load', '_external_load'
 
     def __init__(self, peak_heating: np.ndarray | list = None,
                  peak_cooling: np.ndarray | list = None,
@@ -226,7 +226,8 @@ class Borefield(BaseClass):
         """
         self.borefield = borefield
 
-    def create_rectangular_borefield(self, N_1: int, N_2: int, B_1: int, B_2: int, H: float, D: float = 1, r_b: float = 0.075):
+    def create_rectangular_borefield(self, N_1: int, N_2: int, B_1: int, B_2: int, H: float, D: float = 1,
+                                     r_b: float = 0.075):
         """
         This function creates a rectangular borefield.
         It calls the pygfunction module in the background.
@@ -374,7 +375,7 @@ class Borefield(BaseClass):
         self.custom_gfunction = load_custom_gfunction(location)
         ghe_logger.main_info("Custom g-function has been loaded.")
 
-    def set_investment_cost(self, investment_cost: list =None) -> None:
+    def set_investment_cost(self, investment_cost: list = None) -> None:
         """
         This function sets the investment cost. This is linear with respect to the total field length.
         If None, the default is set.
@@ -480,7 +481,7 @@ class Borefield(BaseClass):
         Load data: LoadData
         """
         return self._borefield_load
-    
+
     @load.setter
     def load(self, load: _LoadData) -> None:
         """
@@ -497,7 +498,7 @@ class Borefield(BaseClass):
         """
         self._borefield_load = load
         self._delete_calculated_temperatures()
-    
+
     @property
     def simulation_period(self) -> int:
         """
@@ -524,7 +525,7 @@ class Borefield(BaseClass):
         None
         """
         self._borefield_load.simulation_period = simulation_period
-        
+
     @property
     def Rb(self) -> float:
         """
@@ -713,6 +714,41 @@ class Borefield(BaseClass):
 
         return self.ground_data.calculate_Tg(H)
 
+    def _check_convergence(self, new_depth: float, old_depth: float, iter: int) -> bool:
+        """
+        This function checks, with the absolute and relative tolerance, if the depth is converged.
+        This is the case if both criteria are met. Raises runtime error if the max number of iterations is crossed.
+
+        Parameters
+        ----------
+        new_depth : float
+            Depth from the current interation
+        old_depth : float
+            Depth from the previous iteration
+        iter : int
+            Current number of iteration
+
+        Returns
+        -------
+        bool
+            True if the depth is converged
+
+        Raises
+        ------
+        RuntimeError
+            RuntimeError if the max number of iterations is crossed
+        """
+        if iter + 1 > self._sizing_setup.max_nb_of_iterations:
+            raise RuntimeError(f'The maximum number of iterations {self._sizing_setup.max_nb_of_iterations} is crossed.'
+                               f'There is no size convergence.')
+        if old_depth == 0:
+            return False
+        test_a_tol = abs(new_depth - old_depth) <= self._sizing_setup.atol if self._sizing_setup.atol != False else True
+        test_rtol = abs(
+            new_depth - old_depth) / old_depth <= self._sizing_setup.rtol if self._sizing_setup.rtol != False else True
+
+        return test_a_tol and test_rtol
+
     def _Ahmadfard(self, th: float, qh: float, qm: float, qa: float) -> float:
         """
         This function sizes the field based on the last year of operation, i.e. quadrants 2 and 4.
@@ -751,7 +787,8 @@ class Borefield(BaseClass):
         time = np.array([th, th + self.load.tm, self.load.ty + self.load.tm + th])
         # Iterates as long as there is no convergence
         # (convergence if difference between depth in iterations is smaller than THRESHOLD_BOREHOLE_DEPTH)
-        while abs(self.H - H_prev) >= Borefield.THRESHOLD_BOREHOLE_DEPTH:
+        i = 0
+        while not self._check_convergence(self.H, H_prev, i):
             # calculate the required g-function values
             gfunct_uniform_T = self.gfunction(time, max(1, self.H))
             # calculate the thermal resistances
@@ -763,6 +800,8 @@ class Borefield(BaseClass):
             # updating the depth values
             H_prev = self.H
             self.H = L / self.number_of_boreholes
+            i += 1
+
         return self.H
 
     def _Carcel(self, th: float, tcm: float, qh: float, qpm: float, qm: float) -> float:
@@ -792,7 +831,8 @@ class Borefield(BaseClass):
 
         # Iterates as long as there is no convergence
         # (convergence if difference between depth in iterations is smaller than THRESHOLD_BOREHOLE_DEPTH)
-        while abs(self.H - H_prev) >= Borefield.THRESHOLD_BOREHOLE_DEPTH:
+        i = 0
+        while not self._check_convergence(self.H, H_prev, i):
             # get the g-function values
             gfunc_uniform_T = self.gfunction(time_steps, max(1, self.H))
 
@@ -807,6 +847,7 @@ class Borefield(BaseClass):
             # updating the depth values
             H_prev = self.H
             self.H = L / self.number_of_boreholes
+            i += 1
         return self.H
 
     def sizing_setup(self, sizing_setup: CalculationSetup = None, use_constant_Rb: bool = None, **kwargs) -> None:
@@ -901,7 +942,8 @@ class Borefield(BaseClass):
         use_constant_Rb_backup = self.borehole.use_constant_Rb
 
         # run the sizing setup
-        self._sizing_setup.update_variables(H_init=H_init, L2_sizing=L2_sizing, L3_sizing=L3_sizing, L4_sizing=L4_sizing,
+        self._sizing_setup.update_variables(H_init=H_init, L2_sizing=L2_sizing, L3_sizing=L3_sizing,
+                                            L4_sizing=L4_sizing,
                                             quadrant_sizing=quadrant_sizing)
         self._sizing_setup.update_variables(**kwargs)
 
@@ -1163,8 +1205,10 @@ class Borefield(BaseClass):
             # size according to a specific quadrant
             self.H = self._size_based_on_temperature_profile(quadrant_sizing, hourly=True)
         else:
-            max_temp = self._size_based_on_temperature_profile(10, hourly=True) if np.any(self.load.hourly_cooling_load) else 0
-            min_temp = self._size_based_on_temperature_profile(20, hourly=True) if np.any(self.load.hourly_heating_load) else 0
+            max_temp = self._size_based_on_temperature_profile(10, hourly=True) if np.any(
+                self.load.hourly_cooling_load) else 0
+            min_temp = self._size_based_on_temperature_profile(20, hourly=True) if np.any(
+                self.load.hourly_heating_load) else 0
             self.H = self._select_size(max_temp, min_temp, True)
             if self.load.imbalance <= 0:
                 # extraction dominated, so quadrants 1 and 4 are relevant
@@ -1217,7 +1261,8 @@ class Borefield(BaseClass):
 
         # Iterates as long as there is no convergence
         # (convergence if difference between depth in iterations is smaller than THRESHOLD_BOREHOLE_DEPTH)
-        while abs(self.H - H_prev) >= Borefield.THRESHOLD_BOREHOLE_DEPTH:
+        i = 0
+        while not self._check_convergence(self.H, H_prev, i):
 
             if hourly:
                 # # calculate g-values
@@ -1301,19 +1346,23 @@ class Borefield(BaseClass):
             if quadrant == 1:
                 # maximum temperature
                 # convert back to required length
-                self.H = (np.max(self.results_peak_cooling[:8760 if hourly else 12]) - self._Tg()) / (self.Tf_max - self._Tg()) * H_prev
+                self.H = (np.max(self.results_peak_cooling[:8760 if hourly else 12]) - self._Tg()) / (
+                        self.Tf_max - self._Tg()) * H_prev
             elif quadrant == 2:
                 # maximum temperature
                 # convert back to required length
-                self.H = (np.max(self.results_peak_cooling[-8760 if hourly else -12:]) - self._Tg()) / (self.Tf_max - self._Tg()) * H_prev
+                self.H = (np.max(self.results_peak_cooling[-8760 if hourly else -12:]) - self._Tg()) / (
+                        self.Tf_max - self._Tg()) * H_prev
             elif quadrant == 3:
                 # minimum temperature
                 # convert back to required length
-                self.H = (np.min(self.results_peak_heating[:8760 if hourly else 12]) - self._Tg()) / (self.Tf_min - self._Tg()) * H_prev
+                self.H = (np.min(self.results_peak_heating[:8760 if hourly else 12]) - self._Tg()) / (
+                        self.Tf_min - self._Tg()) * H_prev
             elif quadrant == 4:
                 # minimum temperature
                 # convert back to required length
-                self.H = (np.min(self.results_peak_heating[-8760 if hourly else -12:]) - self._Tg()) / (self.Tf_min - self._Tg()) * H_prev
+                self.H = (np.min(self.results_peak_heating[-8760 if hourly else -12:]) - self._Tg()) / (
+                        self.Tf_min - self._Tg()) * H_prev
             elif quadrant == 10:
                 # over all years
                 # maximum temperature
@@ -1327,6 +1376,8 @@ class Borefield(BaseClass):
 
             if self.H < 0:
                 return 0
+
+        i += 1
 
         return self.H
 
@@ -1537,13 +1588,15 @@ class Borefield(BaseClass):
             ax.step(time_array, self.results_peak_cooling, 'b-', where="post", lw=1.5, label='Tf peak cooling')
             ax.step(time_array, self.results_peak_heating, 'r-', where="post", lw=1.5, label='Tf peak heating')
 
-            ax.step(time_array, self.results_month_cooling, color='b', linestyle="dashed", where="post", lw=1.5, label='Tf base cooling')
-            ax.step(time_array, self.results_month_heating, color='r', linestyle="dashed", where="post", lw=1.5, label='Tf base heating')
+            ax.step(time_array, self.results_month_cooling, color='b', linestyle="dashed", where="post", lw=1.5,
+                    label='Tf base cooling')
+            ax.step(time_array, self.results_month_heating, color='r', linestyle="dashed", where="post", lw=1.5,
+                    label='Tf base heating')
 
         # define temperature bounds
         ax.hlines(self.Tf_min, 0, self.simulation_period, colors='r', linestyles='dashed', label='', lw=1)
         ax.hlines(self.Tf_max, 0, self.simulation_period, colors='b', linestyles='dashed', label='', lw=1)
-        ax.set_xticks(range(0, self.simulation_period+1, 2))
+        ax.set_xticks(range(0, self.simulation_period + 1, 2))
 
         # Plot legend
         if legend:
@@ -1609,7 +1662,7 @@ class Borefield(BaseClass):
 
             # convolution to get the monthly results
             results = convolve(self.load.monthly_average_load_simulation_period * 1000,
-                               g_value_differences)[:12*self.simulation_period]
+                               g_value_differences)[:12 * self.simulation_period]
 
             # calculation the borehole wall temperature for every month i
             Tb = results / (2 * pi * self.ground_data.k_s) / (H * self.number_of_boreholes) + self._Tg(H)
@@ -1618,15 +1671,19 @@ class Borefield(BaseClass):
             # now the Tf will be calculated based on
             # Tf = Tb + Q * R_b
             results_month_cooling = Tb + self.load.baseload_cooling_power_simulation_period * 1000 \
-                              * (Rb / self.number_of_boreholes / H)
+                                    * (Rb / self.number_of_boreholes / H)
             results_month_heating = Tb - self.load.baseload_heating_power_simulation_period * 1000 \
-                              * (Rb / self.number_of_boreholes / H)
+                                    * (Rb / self.number_of_boreholes / H)
 
             # extra summation if the g-function value for the peak is included
-            results_peak_cooling = results_month_cooling + (self.load.peak_cooling_simulation_period - self.load.baseload_cooling_power_simulation_period) * 1000 \
-                                     * (g_value_peak_cooling / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
-            results_peak_heating = results_month_heating - (self.load.peak_heating_simulation_period - self.load.baseload_heating_power_simulation_period) * 1000 \
-                                   * (g_value_peak_heating / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
+            results_peak_cooling = results_month_cooling + (
+                    self.load.peak_cooling_simulation_period - self.load.baseload_cooling_power_simulation_period) * 1000 \
+                                   * (
+                                           g_value_peak_cooling / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
+            results_peak_heating = results_month_heating - (
+                    self.load.peak_heating_simulation_period - self.load.baseload_heating_power_simulation_period) * 1000 \
+                                   * (
+                                           g_value_peak_heating / self.ground_data.k_s / 2 / pi + Rb) / self.number_of_boreholes / H
 
             # save temperatures under variable
             self.results_peak_heating = results_peak_heating
@@ -1788,12 +1845,12 @@ class Borefield(BaseClass):
         -------
         Reynolds number : float
         """
-        u = self.borehole.fluid_data.mfr / self.borehole.pipe_data.number_of_pipes / self.borehole.fluid_data.rho /\
+        u = self.borehole.fluid_data.mfr / self.borehole.pipe_data.number_of_pipes / self.borehole.fluid_data.rho / \
             (pi * self.borehole.pipe_data.r_in ** 2)
         return self.borehole.fluid_data.rho * u * self.borehole.pipe_data.r_in * 2 / self.borehole.fluid_data.mu
 
-    def optimise_load_profile(self, building_load: HourlyGeothermalLoad, depth: float = None, SCOP: float = 10**6,
-                              SEER: float = 10**6, print_results: bool = False) -> None:
+    def optimise_load_profile(self, building_load: HourlyGeothermalLoad, depth: float = None, SCOP: float = 10 ** 6,
+                              SEER: float = 10 ** 6, print_results: bool = False) -> None:
         """
         This function optimises the load based on the given borefield and the given hourly load.
         (When the load is not geothermal, the SCOP and SEER are used to convert it to a geothermal load.)
@@ -1845,8 +1902,8 @@ class Borefield(BaseClass):
 
         # load hourly heating and cooling load and convert it to geothermal loads
         primary_geothermal_load = HourlyGeothermalLoad()
-        primary_geothermal_load.set_hourly_cooling(building_load.hourly_cooling_load.copy() * (1 + 1/SEER))
-        primary_geothermal_load.set_hourly_heating(building_load.hourly_heating_load.copy() * (1 - 1/SCOP))
+        primary_geothermal_load.set_hourly_cooling(building_load.hourly_cooling_load.copy() * (1 + 1 / SEER))
+        primary_geothermal_load.set_hourly_heating(building_load.hourly_heating_load.copy() * (1 - 1 / SCOP))
 
         # set geothermal load
         self.load = primary_geothermal_load
@@ -1898,8 +1955,8 @@ class Borefield(BaseClass):
 
         # calculate the resulting secundary hourly profile that can be put on the borefield
         self._secundary_borefield_load = HourlyGeothermalLoad()
-        self._secundary_borefield_load.set_hourly_cooling(self.load.hourly_cooling_load / (1 + 1/SEER))
-        self._secundary_borefield_load.set_hourly_heating(self.load.hourly_heating_load / (1 - 1/SCOP))
+        self._secundary_borefield_load.set_hourly_cooling(self.load.hourly_cooling_load / (1 + 1 / SEER))
+        self._secundary_borefield_load.set_hourly_heating(self.load.hourly_heating_load / (1 - 1 / SCOP))
 
         # set building load
         self._building_load = building_load
@@ -1917,7 +1974,8 @@ class Borefield(BaseClass):
 
         if print_results:
             # print results
-            print("The peak load heating is: ", f'{self._secundary_borefield_load.max_peak_heating:.0f}', "kW, leading to",
+            print("The peak load heating is: ", f'{self._secundary_borefield_load.max_peak_heating:.0f}',
+                  "kW, leading to",
                   f'{np.sum(self._secundary_borefield_load.hourly_heating_load):.0f}', "kWh of heating.")
             print("This is", f'{self._percentage_heating:.0f}',
                   "% of the total heating load.")
@@ -1925,7 +1983,8 @@ class Borefield(BaseClass):
                   "kWh of heating should come from another source, with a peak of",
                   f'{self._external_load.max_peak_heating:.0f}', "kW.")
             print("------------------------------------------")
-            print("The peak load cooling is: ", f'{self._secundary_borefield_load.max_peak_cooling:.0f}', "kW, leading to",
+            print("The peak load cooling is: ", f'{self._secundary_borefield_load.max_peak_cooling:.0f}',
+                  "kW, leading to",
                   f'{np.sum(self._secundary_borefield_load.hourly_cooling_load):.0f}', "kWh of cooling.")
             print("This is", f'{self._percentage_cooling:.0f}',
                   "% of the total cooling load.")
@@ -1946,7 +2005,7 @@ class Borefield(BaseClass):
         float
             Percentage of heating load that can be done geothermally.
         """
-        return np.sum(self._secundary_borefield_load.hourly_heating_load) /\
+        return np.sum(self._secundary_borefield_load.hourly_heating_load) / \
             np.sum(self._building_load.hourly_heating_load) * 100
 
     @property
@@ -1959,7 +2018,7 @@ class Borefield(BaseClass):
         float
             Percentage of cooling load that can be done geothermally.
         """
-        return np.sum(self._secundary_borefield_load.hourly_cooling_load) /\
+        return np.sum(self._secundary_borefield_load.hourly_cooling_load) / \
             np.sum(self._building_load.hourly_cooling_load) * 100
 
     def calculate_quadrant(self) -> int:
@@ -2084,8 +2143,12 @@ class Borefield(BaseClass):
         for i in range(self.borehole.pipe_data.number_of_pipes):
             circles_outer.append(plt.Circle(pos[i], self.borehole.pipe_data.r_out, color="black"))
             circles_inner.append(plt.Circle(pos[i], self.borehole.pipe_data.r_in, color="red"))
-            circles_outer.append(plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_out, color="black"))
-            circles_inner.append(plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_in, color="blue"))
+            circles_outer.append(
+                plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_out,
+                           color="black"))
+            circles_inner.append(
+                plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_in,
+                           color="blue"))
 
         # set visual settings for figure
         axes.set_aspect('equal')